Unify application entrypoint
