+++ Procfile
@@ -1,1 +1,1 @@
-web: python main.py
+web: gunicorn --bind 0.0.0.0:${PORT:-5000} main:app
